+++ deps.ts
@@ -5,8 +5,8 @@ export * as hash from "https://deno.land/std@0.65.0/hash/mod.ts";
 // @deno-types="https://unpkg.com/cac@6.6.1/mod.d.ts"
 export { cac } from "https://unpkg.com/cac@6.6.1/mod.js";
 
-// @deno-types="https://unpkg.com/rollup@2.23.0/dist/rollup.d.ts"
-export { rollup } from "https://unpkg.com/rollup@2.23.0/dist/es/rollup.browser.js";
+// @deno-types="https://unpkg.com/rollup@2.26.3/dist/rollup.d.ts"
+export { rollup } from "https://unpkg.com/rollup@2.26.3/dist/es/rollup.browser.js";
 export type {
   Plugin,
   ResolveIdResult,
@@ -15,9 +15,11 @@ export type {
   RollupBuild,
   RollupCache,
   OutputAsset,
+  OutputBundle,
   OutputChunk,
   OutputOptions,
-} from "https://unpkg.com/rollup@2.23.0/dist/rollup.d.ts";
+  ModuleFormat,
+} from "https://unpkg.com/rollup@2.26.3/dist/rollup.d.ts";
 
 import { default as _debounce } from "https://unpkg.com/lodash-es@4.17.15/debounce.js";