+++ plugin/mod.ts
@@ -1,10 +1,11 @@
 export * from "./hooks.ts";
 export * from "./importResolver/mod.ts";
 export * from "./rootResolver/mod.ts";
-export * from "./cacheResolver/mod.ts";
 export * from "./chainResolver/mod.ts";
 export { pluginCacheLoader, Opts as CacheLoaderOptions } from "./cacheLoader/mod.ts";
 export { pluginFileLoader, Opts as FileLoaderOptions } from "./fileLoader/mod.ts";
 export * from "./terserTransform/mod.ts";
 export * from "./typescriptTransform/mod.ts";
 export * from "./typescriptCompile/mod.ts";
+export * from "./cssBundle/mod.ts";
+export * from "./htmlBundle/mod.ts";