+++ recipes/react-template/src/deps.ts
@@ -1,0 +1,6 @@
+// @deno-type="https://cdn.jsdelivr.net/gh/soremwar/deno_types@master/react-dom/v16.13.1/react-dom.d.ts";
+import ReactDOM from "https://dev.jspm.io/react-dom@16.13.1";
+// @deno-types="https://cdn.jsdelivr.net/gh/soremwar/deno_types@master/react/v16.13.1/react.d.ts"
+import React from "https://dev.jspm.io/react@16.13.1";
+
+export { React, ReactDOM };