+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Plugins for html and css assets
+- React template recipe
+- Plugins are now also exported with `export default`
+
+### Changed
+
+- Updated Rollup to v2.26.3
+- Toplevel mod.ts now re-exports rollup from deps.ts
+- Removed pluginCacheResolver
+
 ## [0.5.0](https://github.com/denofn/denopack/compare/0.4.0..0.5.0) - 2020-08-13
 
 ### Changed